Change modules structure for efficiency. Changed requirements for index css file name
