+++ README.md
@@ -1,2 +1,28 @@
-# Awesome-Books-with-ES6
- Awesome Books Application with ES6 syntax
+# Awesome Books ES-6
+ ES-6 Awesome Books web-page. Library.
+
+## Built With
+ Github, Visual Studio Code, HTML, CSS & Javascript
+
+## You can find us in:
+ [Live Demo](https://sntgopl.github.io/Awesome-Books-ES6/)
+
+## Getting started
+ Welcome to my new website. Designed specially for creating a library with all your favorite books and most-wanted too. Sharing and comparing your libraries with your friends. We have used Javascript ES-6 syntax and modules for creating dynamic functions for more addaptability.
+
+## Author
+ Github: @sntgopl
+ Twitter: @LezicaPons
+ Linkedin: Santiago Pons Lezica
+
+## Contribution
+ Contributions, issues and reviews are very much welcomed.
+
+## Support
+ Please feel free to rate this project.
+
+## Acknowledgements 
+ Microverse Community and Github Repositories.
+
+## License
+ This project is Me Licensed.
