+++ modules/classes.js
@@ -1,0 +1,18 @@
+import { booksArray } from './data&storage.js';
+
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+
+  addBook() {
+    booksArray.push(this);
+  }
+
+  static removeBooks(index) {
+    booksArray.splice(index, 1);
+  }
+}
+
+export default Book;
